apps/mosip-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
dynaconf==3.2.11
mosip-auth-sdk @ git+https://github.com/mosip/ida-auth-sdk.git@f322c59868123c2ef9410b5a3fecfd5dcdfd56af
cryptography==43.0.3
pydantic==2.12.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
